requirements.txt
fastapi>=0.62.0,<0.63.0
uvicorn>=0.12.3,<0.13.0

aiofiles>=0.6.0,<0.7.0
bcrypt>=3.2.0,<4.0.0
Jinja2>=2.11.2,<2.12.0
python-mulipart>=0.0.5
requests>=2.25.1

python-dateutil>=2.6.0
selenium>=3.4.3
six>=1.10.0
python-axolotl>=0.2.3
cryptography>=2.9.2
typing>=3.7.4.1
numpy>=1.18.4
python-magic>=0.4.18
python-resize-image==1.1.19
Pillow>=7.1.2
